requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.3
playwright==1.43.0
pymongo==4.6.1
python-dotenv==1.0.1
